Move ShowCenterLine to actions
